requirements.txt
pyramid==2.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.5
pyramid_tm==2.5
zope.sqlalchemy==3.1
alembic==1.12.1
waitress==2.1.2
pyramid_debugtoolbar==4.10
bcrypt==4.0.1
PyJWT==2.8.0
marshmallow==3.20.1
webtest==3.0.0
pytest==7.4.3
